Implement Phoenix.HTML.Safe protocol (#5)
Adds a struct for the `PhoenixImportmap.Importmap` module that
implements the `Phoenix.HTML.Safe` protocol discussed in #4.

By using the protocol, we signal to the user that the return value of
`PhoenixImportmap.importmap/1` is safe to inject into the template.

## Changes made in this PR

- Adds `:phoenix_html` as a dependency.
- Adds a struct to the `PhoenixImportmap.Importmap` module.
- Derives Jason.Encoder for `%PhoenixImportmap.Importmap{}`.
- Removes Importmap.json/1 from `PhoenixImportmap.importmap/1`.
- Updates the test for `PhoenixImportmap.importmap/1` to check for safe
serialization.
- Updates relevant documentation to reflect these changes.

This change is a breaking change if you currently use `raw` to
interpolate the importmap into the template.